sfpmgr: send indication of whether RX CDR is present to consumers
Send an indication of whether an RX CDR/retimer is present in the
module, since the system may need to know this to determine how
exactly to program the port.

There is a similar indication for SFPs defined in SFF-8472 (byte 64,
bit 3), but it isn't specific to RX/TX and therefore isn't of as much
use and so not sent.

VRVDR-50555