vyatta-sfpd: allow all SFP helpers to trigger insert/remove from command
Move the implementation of process_sfpinsertedremoved into
BaseSfpHelper so that all SFP helpers get it which allows injection of
commands to trigger SFP insertion and removal without a hardware
event, which is useful for unit testing. Extend the command to allow
the passing in of the port type (SFP or QSFP) for full flexibility of
testing across platforms, but defaulting to SFP for backwards
compatibility.

Extend the command also to allow the passing of the extra_state that
will be published to the FAL plugin, for full control of the
information for unit testing.

Exception handling is relied upon for handling errors as per Python
convention, rather than a bool result.

VRVDR-51235