marvell88e1111: modify registers instead of setting during SGMII enable
Rather than setting whole registers which may include writing to
reserved bits, and making it unclear which bits are being set, instead
introduce a helper function that modifies a register and use this in
enable_sgmii.

In addition, make this helper function do the necessary endian swaps
so that the values passed into this function match up with the bit
positions in datasheet.