marvell88e1111: modify registers instead of setting during AN caps setting
Rather than setting whole registers which may include writing to
reserved bits, and making it unclear which bits are being set, instead
modify registers in set_autoneg_caps and use symbols for the values.

Also make use of the same symbols in get_linkpartner_caps,
endian-swapping the read of the register to match the endianness of
the register values.
In addition, make this helper function do the necessary endian swaps
so that the values passed into this function match up with the bit
positions in datasheet.