marvell88e1111: wait for reset to complete during setting of AN caps
Wait for the reset to complete during setting of autoneg caps to avoid
writes being ignored, and split the enabling of autoneg into a prior
register write since it's not clear what the PHY does when this is set
at the same time as the bit to initiate the soft reset.
This therefore ensures that further writes have the desired effect.

VRVDR-50921