marvell88e1111: wait for reset to complete during force speed setting
Wait for the reset to complete during forced speed setting and turning
off auto-negotiate to avoid subsequent writes being ignored, and split
reset operation out since it's not clear what the PHY does when other
things are set at the same time as the bit to initiate the soft reset.

This therefore ensures that further writes have the desired effect.

VRVDR-50921